add Upstart 1.8 job files
:Release Notes:
This fix is simply moving the Upstart 1.8 job files for db8
from the starfish-initscripts package to their proper home.

:Detailed Notes:
As mentioned above, this fix is just moving the upstart script - in the
initial implementation of Upstart 1.8, many 'component' upstart jobs
were modified for 1.8 and packaged with starfish-initscripts to ease
the transition from Upstart 0.3.11 to 1.8.  The plan was and is to
move the 'imported' scripts back to their home package as soon as is
reasonable.

:Testing Performed:
Since the file itself is functionally unchanged from that present in
starfish-initscripts, the testing done running Upstart 1.8 + s-i over
the last couple of months still applies.

The testing done specifically for this fix is to locally build the
component and inspect the output directory:
starfish/BUILD-goldfinger/work/armv7a-vfp-neon-starfish-linux-gnueabi/ \
   /db8/2.0.0-72-r13starfish1/image/etc/init and ..../etc/event.d
to confirm that the properly configured upstart scripts end up in their
respective locations.

:QA Notes:

:Issues Addressed:
[GF-5620]  update coreOS upstart scripts for 1.8

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: I68b57689a7a4f80cc7bfbb036306fd584c675d3e
Reviewed-on: https://g2g.palm.com/2712
Reviewed-by: Build Verification
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Roger Stringer <roger.stringer@lge.com>