Add parameter to query for inactive records
:Release Notes:
Add a new element to the MojDbQuery class, m_ignoreInactiveShards with
a default value of true. Add methods to allow it to be set and read.
Add equivalent element to MojDbIsamQuery class initialized from the
related query class.
Wrap the logic to exclude records for inactive shards in a condition
that MojDbIsamQuery::m_ignoreInactiveShards is true.

:Detailed Notes:
New methods in MojDbQuery are:
- setIgnoreInactiveShards(bool val = true)
- ignoreInactiveShards()

:Testing Performed:
Needs [GF-38548] to test. Can't be independently tested without
implementing that or something similar to it.

:QA Notes:

:Issues Addressed:
[GF-38358] DB8 query needs option to pass through records for inactive shards

Open-webOS-DCO-1.0-Signed-off-by: Roger Stringer <roger.stringer@lge.com>

Change-Id: I6e9974a52df1104787ebeccb4bfcb741b4a7a5f7
Reviewed-on: https://g2g.palm.com/3453
Reviewed-by: Nikolay Orliuk <nikolay.orliuk@lge.com>
Tested-by: Nikolay Orliuk <nikolay.orliuk@lge.com>
Reviewed-by: Maksym Sditanov <maxim.sditanov@lge.com>
Reviewed-by: Build Verification
Reviewed-by: Roger Stringer <roger.stringer@lge.com>