Increment DB8 version from 3.0 to 3.1
:Release Notes:
Implementation of [GF-37501] adds a new filter operator to DB8, which is
a non-trivial change to the API and applications/services using DB8 which
use the new "%%" operator will need to ensure they are using the correct
version of DB8.
This also includes the change the 'shardInfo' API which returns two
parameter values correctly as a boolean instead of string.

:Detailed Notes:

:Testing Performed:
unit tests
mini-BAT

:QA Notes:

:Issues Addressed:

Open-webOS-DCO-1.0-Signed-off-by: Roger Stringer <roger.stringer@lge.com>

Change-Id: Ib82db477b4dffa697f18f80aba05ed54f71cf205
Reviewed-on: https://g2g.palm.com/3477
Tested-by: Roger Stringer <roger.stringer@lge.com>
Reviewed-by: Build Verification
Reviewed-by: Maksym Sditanov <maxim.sditanov@lge.com>
Reviewed-by: Roger Stringer <roger.stringer@lge.com>